rename sdk related
